pyproject.toml
[project]
name = "back-to-back"
version = "0.0.0"
description = "Add your description here"
readme = "README.md"
license = {text = "MIT"}
authors = [
    { name = "Your Name", email = "your.name@example.com" }
]
maintainers = [
    { name = "Your Name", email = "your.name@example.com" }
]
requires-python = ">=3.13"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3 :: Only",
    "Typing :: Typed",
]
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "openai>=1.27.0",
    "pipecat-ai[openai,deepgram,flows]==0.0.76",
    "pydantic>=2.10.6,<2.11.dev0",
    "redis>=6.2.0",
    "uvicorn>=0.35.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "ruff>=0.1",
]

[project.scripts]
back-to-back = "back_to_back.main:main"

[build-system]
requires = ["hatchling", "editables"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py313"
line-length = 100
extend-include = ["*.pyi"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
    "PL",   # pylint
    "PTH",  # flake8-use-pathlib
]
ignore = ["PLR0913"]  # Too many arguments

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_equality = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = [
    "-ra",
    "--strict-markers",
    "--cov=back_to_back",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["src/back_to_back"]
omit = ["*/__main__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.hatch.build.targets.wheel]
packages = ["src/back_to_back"]
